python/kafka-consumer-groups-sync/src/requirements.txt
pyyaml
kafka-python
lz4
crc32c